+++ CHANGES.md
@@ -52,6 +52,7 @@
 * Update Requests library 2.7.0 (ab1f493) to 2.7.0 (8b5e457)
 * Update Tornado webserver from 4.2.dev1 (609dbb9) to 4.2b1 (61a16c9)
 * Change reload_module call to explicit import lib/six.moves
+* Change queue, httplib, cookielib and xmlrpclib to use explicit import lib/six.moves
 
 
 ### 0.9.1 (2015-05-25 03:03:00 UTC)